backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy[asyncio]>=2.0.0
python-dotenv>=1.0.0
meilisearch-python-sdk>=1.0.0
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
httpx>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.28.0
greenlet>=2.0.0
